+++ lib/detroit/core_ext/basic_object.rb
@@ -1,6 +1,0 @@
-class BasicObject
-
-  def self.preinitialize
-  end
-
-end